+++ src/db/mod.rs
@@ -1,2 +1,3 @@
+pub mod alarm;
 pub mod online;
 pub mod sensor;