+++ src/tests_integration/mod.rs
@@ -1,1 +1,2 @@
+mod alarm_tests;
 mod replay_tests;